+++ SCons/SConf.py
@@ -43,7 +43,7 @@ import traceback
 import SCons.Action
 import SCons.Builder
 import SCons.Errors
-import SCons.Job
+import SCons.Taskmaster.Job
 import SCons.Node.FS
 import SCons.Taskmaster
 import SCons.Util
@@ -551,7 +551,7 @@ class SConfBase:
             SConfFS.set_max_drift(0)
             tm = SCons.Taskmaster.Taskmaster(nodes, SConfBuildTask)
             # we don't want to build tests in parallel
-            jobs = SCons.Job.Jobs(1, tm )
+            jobs = SCons.Taskmaster.Job.Jobs(1, tm)
             jobs.run()
             for n in nodes:
                 state = n.get_state()
